+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add profile option to Memplify.report for environment control
+
 ## [0.1.0] - 2025-03-02
 
 - Initial release